seguimos peleando :(

+++ database/migrations/2021_03_07_222921_create_usuarios_table.php
@@ -14,7 +14,7 @@ class CreateUsuariosTable extends Migration
     public function up()
     {
         Schema::create('usuarios', function (Blueprint $table) {
-            $table->integer('cuil')->unsigned()->primary();
+            $table->bigInteger('cuil')->unsigned()->primary();
             $table->string('nombre');
             $table->string('apellido');
             $table->string('direccion');
