+++ database/migrations/2021_03_07_225035_create_donacion_1_table.php
@@ -15,19 +15,19 @@ class CreateDonacion1Table extends Migration
     {
         Schema::create('donacion_1', function (Blueprint $table) {
             $table->integer('id_proceso')->unsigned()->primary();
-            $table->integer('cuil_ablacionista')->unsigned();
+            $table->bigInteger('cuil_ablacionista')->unsigned();
             $table->foreign('cuil_ablacionista')->references('cuil')->on('usuarios');
             $table->boolean('cert_defuncion')->nullable();
-            $table->integer('id_provincia')->unsigned();
+            $table->bigInteger('id_provincia')->unsigned();
             $table->foreign('id_provincia')->references('id_prov')->on('provincias');
-            $table->integer('contraindicaciones')->unsigned();
+            $table->bigInteger('contraindicaciones')->unsigned();
             $table->foreign('contraindicaciones')->references('id_contraindicacion')->on('contraindicaciones');
             $table->string('manifestacion_vol')->nullable();
             $table->boolean('viabilidad')->nullable();
             $table->boolean('coordinacion')->nullable();
             $table->boolean('deteccion')->nullable();
             $table->boolean('ev_hc_globo')->nullable();
-            $table->integer('id_inst_donante')->unsigned();
+            $table->bigInteger('id_inst_donante')->unsigned();
             $table->foreign('id_inst_donante')->references('id_institucion')->on('institucion_donante');
 
         });
